Initial commit. Initial simple physics simulation and track generation
With this commit, a working car physics simulation is setup.
 - Cars drive based on a simple car-controller rules.
 - Cars collide with track.
 - Cars collide with each other.

A track generator is created. This will read a .png image and create a
track from it. The png file is drawn with black background, and the
drivable area is drawn as white. The program will extract the black->white
transition and setup rectangles on perimeter of the track to enable simple
collisions with the track.

Parameters such as the track png-image, starting position of cars,
physical constants are currently all hard-coded.

The commit also introduce maven build scripts and some simple instructions
on how to build the project with maven in the Readme.md file.